+++ gridkit-rs/src/data_tile.rs
@@ -126,13 +126,12 @@ impl<
         let dx = (bounds.2 - bounds.0) / nx as f64;
         let dy = (bounds.3 - bounds.1) / ny as f64;
         let mut grid = RectGrid::new(dx, dy);
-        // nocheckin determine offset!
         let start_cell_centroid = [bounds.0 + dx/2., bounds.1+dy/2.];
         grid.anchor_inplace(&start_cell_centroid, CellElement::Centroid);
         let start_id = grid.cell_at_point(&start_cell_centroid);
         let tile = Tile {
             grid: Grid::RectGrid(grid),
-            start_id: start_id.into(), // nocheckin, align tpye of start_id, use [] intead of ()
+            start_id: start_id.into(), // TODO: align tpye of start_id, use [] intead of ()
             nx,
             ny,
         };